+++ src/util.rs
@@ -1,0 +1,21 @@
+pub fn hsv_to_rgb(h: f32, s: f32, v: f32) -> (u8, u8, u8) {
+    let i = (h * 6.0).floor() as i32;
+    let f = h.mul_add(6.0, -(i as f32));
+    let p = v * (1.0 - s);
+    let q = v * f.mul_add(-s, 1.0);
+    let t = v * (1.0 - f).mul_add(-s, 1.0);
+    let (r, g, b) = match i.rem_euclid(6) {
+        0 => (v, t, p),
+        1 => (q, v, p),
+        2 => (p, v, t),
+        3 => (p, q, v),
+        4 => (t, p, v),
+        5 => (v, p, q),
+        _ => (0.0, 0.0, 0.0),
+    };
+    (
+        (r * 255.0).round() as u8,
+        (g * 255.0).round() as u8,
+        (b * 255.0).round() as u8,
+    )
+}
